+++ extensions/say/src/tools/say.ts
@@ -3,6 +3,10 @@ import { getParsedSaySettings } from "@/utils";
 
 type Input = {
   /**
+   * The voice to use. If not provided, the default voice will be used.
+   */
+  voice?: string;
+  /**
    * The text to say
    */
   content: string;
@@ -11,8 +15,9 @@ type Input = {
 /**
  * Use AI to say things out loud
  */
-export default async function ({ content }: Input) {
+export default async function ({ content, voice }: Input) {
   const { keepSilentOnError, ...saySettings } = getParsedSaySettings();
+  saySettings.voice = voice ?? saySettings.voice;
   try {
     await say(content, saySettings);
   } catch (error) {